added better status ui front-end for decisions
